finito

+++ database/migrations/2022_07_18_135118_create_comics_table.php
@@ -18,7 +18,7 @@ class CreateComicsTable extends Migration
             $table->string('title');
             $table->string('description');
             $table->string('thumb');
-            $table->float('price', 4, 2);
+            $table->integer('price');
             $table->string('series');
             $table->date('sale_date');
             $table->string('type');
